pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "2D-game-editor"
version = "0.1.0"
description = "A modular and extensible 2D game editor built with Python and Pygame."
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
readme = "README.md"
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Operating System :: OS Independent",
]
dependencies = [
    "pygame>=2.0",
    "Pillow>=8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "flake8>=3.9",
    "black>=21.0",
    "mypy>=0.9",
]

[project.urls]
Homepage = "https://github.com/yourusername/2D-game-editor"
Documentation = "https://github.com/yourusername/2D-game-editor/wiki"
Repository = "https://github.com/yourusername/2D-game-editor"
Issues = "https://github.com/yourusername/2D-game-editor/issues"

[tool.setuptools]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.tox
  | venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
python_classes = "Test*"
addopts = "--verbose --cov=src --cov-report=term-missing"
